+++ docs/changelog.rst
@@ -74,8 +74,10 @@ consumption to do the same tasks.
 Detailed list of changes
 -------------------------------------
 
-0.36.3 [future]
+0.36.3 [2024-09-25]
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+- The option ``second_transparent_bg`` has been removed and replaced by :opt:`transparent_background_colors` which allows setting up to seven additional colors that will be transparent, with individual opacities per color (:iss:`7646`)
 
 - Fix a regression in the previous release that broke use of the ``cd`` command in session files (:iss:`7829`)
 
@@ -136,7 +138,7 @@ Detailed list of changes
 
 - launch command: A new :option:`launch --bias` option to adjust the size of newly created windows declaratively (:iss:`7634`)
 
-- A new option :opt:`second_transparent_bg` to make a second background color semi-transparent via :opt:`background_opacity`. Useful for things like cursor line highlight in editors (:iss:`7646`)
+- A new option :opt:`transparent_background_colors` to make a second background color semi-transparent via :opt:`background_opacity`. Useful for things like cursor line highlight in editors (:iss:`7646`)
 
 - A new :doc:`notify </kittens/notify>` kitten to show desktop notifications
   from the command line with support for icons, buttons and more.